Properly organise event-firing/internal hooks
