+++ patches/minecraft/net/minecraft/block/BlockChorusFlower.java.patch
@@ -5,7 +5,7 @@
           if (p_196267_2_.func_175623_d(blockpos) && blockpos.func_177956_o() < 256) {
              int i = p_196267_1_.func_177229_b(field_185607_a);
 -            if (i < 5) {
-+            if (i < 5 && net.minecraftforge.common.ForgeHooks.onCropsGrowPre(p_196267_2_, blockpos, p_196267_1_, true)) {
++            if (i < 5 && net.minecraftforge.event.ForgeEventFactory.onCropsGrowPre(p_196267_2_, blockpos, p_196267_1_, true)) {
                 boolean flag = false;
                 boolean flag1 = false;
                 IBlockState iblockstate = p_196267_2_.func_180495_p(p_196267_3_.func_177977_b());
@@ -14,7 +14,7 @@
                    this.func_185605_c(p_196267_2_, p_196267_3_);
                 }
 -
-+               net.minecraftforge.common.ForgeHooks.onCropsGrowPost(p_196267_2_, p_196267_3_, p_196267_1_);
++               net.minecraftforge.event.ForgeEventFactory.onCropsGrowPost(p_196267_2_, p_196267_3_, p_196267_1_);
              }
           }
        }
