+++ patches/minecraft/net/minecraft/block/BlockCrops.java.patch
@@ -10,9 +10,9 @@
           if (i < this.func_185526_g()) {
              float f = func_180672_a(this, p_196267_2_, p_196267_3_);
 -            if (p_196267_4_.nextInt((int)(25.0F / f) + 1) == 0) {
-+            if (net.minecraftforge.common.ForgeHooks.onCropsGrowPre(p_196267_2_, p_196267_3_, p_196267_1_, p_196267_4_.nextInt((int)(25.0F / f) + 1) == 0)) {
++            if (net.minecraftforge.event.ForgeEventFactory.onCropsGrowPre(p_196267_2_, p_196267_3_, p_196267_1_, p_196267_4_.nextInt((int)(25.0F / f) + 1) == 0)) {
                 p_196267_2_.func_180501_a(p_196267_3_, this.func_185528_e(i + 1), 2);
-+               net.minecraftforge.common.ForgeHooks.onCropsGrowPost(p_196267_2_, p_196267_3_, p_196267_1_);
++               net.minecraftforge.event.ForgeEventFactory.onCropsGrowPost(p_196267_2_, p_196267_3_, p_196267_1_);
              }
           }
        }
