+++ patches/minecraft/net/minecraft/block/BlockReed.java.patch
@@ -13,14 +13,14 @@
  
           if (i < 3) {
              int j = p_196267_1_.func_177229_b(field_176355_a);
-+            if (net.minecraftforge.common.ForgeHooks.onCropsGrowPre(p_196267_2_, p_196267_3_, p_196267_1_, true)) {
++            if (net.minecraftforge.event.ForgeEventFactory.onCropsGrowPre(p_196267_2_, p_196267_3_, p_196267_1_, true)) {
              if (j == 15) {
                 p_196267_2_.func_175656_a(p_196267_3_.func_177984_a(), this.func_176223_P());
                 p_196267_2_.func_180501_a(p_196267_3_, p_196267_1_.func_206870_a(field_176355_a, Integer.valueOf(0)), 4);
              } else {
                 p_196267_2_.func_180501_a(p_196267_3_, p_196267_1_.func_206870_a(field_176355_a, Integer.valueOf(j + 1)), 4);
              }
-+            net.minecraftforge.common.ForgeHooks.onCropsGrowPost(p_196267_2_, p_196267_3_, p_196267_1_);
++            net.minecraftforge.event.ForgeEventFactory.onCropsGrowPost(p_196267_2_, p_196267_3_, p_196267_1_);
 +            }
           }
        }
