+++ patches/minecraft/net/minecraft/block/BlockStem.java.patch
@@ -8,7 +8,7 @@
        if (p_196267_2_.func_201669_a(p_196267_3_.func_177984_a(), 0) >= 9) {
           float f = BlockCrops.func_180672_a(this, p_196267_2_, p_196267_3_);
 -         if (p_196267_4_.nextInt((int)(25.0F / f) + 1) == 0) {
-+         if (net.minecraftforge.common.ForgeHooks.onCropsGrowPre(p_196267_2_, p_196267_3_, p_196267_1_, p_196267_4_.nextInt((int)(25.0F / f) + 1) == 0)) {
++         if (net.minecraftforge.event.ForgeEventFactory.onCropsGrowPre(p_196267_2_, p_196267_3_, p_196267_1_, p_196267_4_.nextInt((int)(25.0F / f) + 1) == 0)) {
              int i = p_196267_1_.func_177229_b(field_176484_a);
              if (i < 7) {
 -               p_196267_1_ = p_196267_1_.func_206870_a(field_176484_a, Integer.valueOf(i + 1));
@@ -26,7 +26,7 @@
                    p_196267_2_.func_175656_a(p_196267_3_, this.field_149877_a.func_196523_e().func_176223_P().func_206870_a(BlockHorizontal.field_185512_D, enumfacing));
                 }
              }
-+            net.minecraftforge.common.ForgeHooks.onCropsGrowPost(p_196267_2_, p_196267_3_, p_196267_1_);
++            net.minecraftforge.event.ForgeEventFactory.onCropsGrowPost(p_196267_2_, p_196267_3_, p_196267_1_);
           }
  
        }
