+++ patches/minecraft/net/minecraft/client/KeyboardListener.java.patch
@@ -24,15 +24,15 @@
                 if (p_197961_5_ != 1 && (p_197961_5_ != 2 || !this.field_197973_b)) {
                    if (p_197961_5_ == 0) {
 -                     aboolean[0] = iguieventlistener.keyReleased(p_197961_3_, p_197961_4_, p_197961_6_);
-+                     aboolean[0] = net.minecraftforge.client.ForgeHooksClient.onGuiKeyReleasedPre(this.field_197972_a.field_71462_r, p_197961_3_, p_197961_4_, p_197961_6_);
++                     aboolean[0] = net.minecraftforge.client.event.ForgeClientEventFactory.onGuiKeyReleasedPre(this.field_197972_a.field_71462_r, p_197961_3_, p_197961_4_, p_197961_6_);
 +                     if (!aboolean[0]) aboolean[0] = iguieventlistener.keyReleased(p_197961_3_, p_197961_4_, p_197961_6_);
-+                     if (!aboolean[0]) aboolean[0] = net.minecraftforge.client.ForgeHooksClient.onGuiKeyReleasedPost(this.field_197972_a.field_71462_r, p_197961_3_, p_197961_4_, p_197961_6_);
++                     if (!aboolean[0]) aboolean[0] = net.minecraftforge.client.event.ForgeClientEventFactory.onGuiKeyReleasedPost(this.field_197972_a.field_71462_r, p_197961_3_, p_197961_4_, p_197961_6_);
                    }
                 } else {
 -                  aboolean[0] = iguieventlistener.keyPressed(p_197961_3_, p_197961_4_, p_197961_6_);
-+                  aboolean[0] = net.minecraftforge.client.ForgeHooksClient.onGuiKeyPressedPre(this.field_197972_a.field_71462_r, p_197961_3_, p_197961_4_, p_197961_6_);
++                  aboolean[0] = net.minecraftforge.client.event.ForgeClientEventFactory.onGuiKeyPressedPre(this.field_197972_a.field_71462_r, p_197961_3_, p_197961_4_, p_197961_6_);
 +                  if (!aboolean[0]) aboolean[0] = iguieventlistener.keyPressed(p_197961_3_, p_197961_4_, p_197961_6_);
-+                  if (!aboolean[0]) aboolean[0] = net.minecraftforge.client.ForgeHooksClient.onGuiKeyPressedPost(this.field_197972_a.field_71462_r, p_197961_3_, p_197961_4_, p_197961_6_);
++                  if (!aboolean[0]) aboolean[0] = net.minecraftforge.client.event.ForgeClientEventFactory.onGuiKeyPressedPost(this.field_197972_a.field_71462_r, p_197961_3_, p_197961_4_, p_197961_6_);
                 }
  
              }, "keyPressed event handler", iguieventlistener.getClass().getCanonicalName());
@@ -41,7 +41,7 @@
              }
           }
 -
-+         net.minecraftforge.client.ForgeHooksClient.fireKeyInput(p_197961_3_, p_197961_4_, p_197961_5_, p_197961_6_);
++         net.minecraftforge.client.event.ForgeClientEventFactory.fireKeyInput(p_197961_3_, p_197961_4_, p_197961_5_, p_197961_6_);
        }
     }
  
@@ -50,17 +50,17 @@
              if (Character.charCount(p_197963_3_) == 1) {
                 GuiScreen.func_195121_a(() -> {
 -                  iguieventlistener.charTyped((char)p_197963_3_, p_197963_4_);
-+                  if (net.minecraftforge.client.ForgeHooksClient.onGuiCharTypedPre(this.field_197972_a.field_71462_r, (char)p_197963_3_, p_197963_4_)) return;
++                  if (net.minecraftforge.client.event.ForgeClientEventFactory.onGuiCharTypedPre(this.field_197972_a.field_71462_r, (char)p_197963_3_, p_197963_4_)) return;
 +                  if (iguieventlistener.charTyped((char)p_197963_3_, p_197963_4_)) return;
-+                  net.minecraftforge.client.ForgeHooksClient.onGuiCharTypedPost(this.field_197972_a.field_71462_r, (char)p_197963_3_, p_197963_4_);
++                  net.minecraftforge.client.event.ForgeClientEventFactory.onGuiCharTypedPost(this.field_197972_a.field_71462_r, (char)p_197963_3_, p_197963_4_);
                 }, "charTyped event handler", iguieventlistener.getClass().getCanonicalName());
              } else {
                 for(char c0 : Character.toChars(p_197963_3_)) {
                    GuiScreen.func_195121_a(() -> {
 -                     iguieventlistener.charTyped(c0, p_197963_4_);
-+                     if (net.minecraftforge.client.ForgeHooksClient.onGuiCharTypedPre(this.field_197972_a.field_71462_r, c0, p_197963_4_)) return;
++                     if (net.minecraftforge.client.event.ForgeClientEventFactory.onGuiCharTypedPre(this.field_197972_a.field_71462_r, c0, p_197963_4_)) return;
 +                     if (iguieventlistener.charTyped(c0, p_197963_4_)) return;
-+                     net.minecraftforge.client.ForgeHooksClient.onGuiCharTypedPost(this.field_197972_a.field_71462_r, c0, p_197963_4_);
++                     net.minecraftforge.client.event.ForgeClientEventFactory.onGuiCharTypedPost(this.field_197972_a.field_71462_r, c0, p_197963_4_);
                    }, "charTyped event handler", iguieventlistener.getClass().getCanonicalName());
                 }
              }
