+++ patches/minecraft/net/minecraft/client/audio/SoundManager.java.patch
@@ -20,7 +20,7 @@
  
     public void func_148611_c(ISound p_148611_1_) {
        if (this.field_148617_f) {
-+         p_148611_1_ = net.minecraftforge.client.ForgeHooksClient.playSound(this, p_148611_1_);
++         p_148611_1_ = net.minecraftforge.client.event.ForgeClientEventFactory.playSound(this, p_148611_1_);
 +         if (p_148611_1_ == null) return;
           SoundEventAccessor soundeventaccessor = p_148611_1_.func_184366_a(this.field_148622_c);
           ResourceLocation resourcelocation = p_148611_1_.func_147650_b();
