+++ patches/minecraft/net/minecraft/client/entity/AbstractClientPlayer.java.patch
@@ -14,6 +14,6 @@
        }
  
 -      return f;
-+      return net.minecraftforge.client.ForgeHooksClient.getOffsetFOV(this, f);
++      return net.minecraftforge.client.event.ForgeClientEventFactory.getOffsetFOV(this, f);
     }
  }
