+++ patches/minecraft/net/minecraft/client/entity/EntityOtherPlayerMP.java.patch
@@ -4,7 +4,7 @@
     }
  
     public boolean func_70097_a(DamageSource p_70097_1_, float p_70097_2_) {
-+      net.minecraftforge.common.ForgeHooks.onPlayerAttack(this, p_70097_1_, p_70097_2_);
++      net.minecraftforge.event.ForgeEventFactory.onPlayerAttack(this, p_70097_1_, p_70097_2_);
        return true;
     }
