+++ patches/minecraft/net/minecraft/client/gui/GuiBossOverlay.java.patch
@@ -5,7 +5,7 @@
           for(BossInfoClient bossinfoclient : this.field_184060_g.values()) {
              int k = i / 2 - 91;
 +            net.minecraftforge.client.event.RenderGameOverlayEvent.BossInfo event =
-+                    net.minecraftforge.client.ForgeHooksClient.bossBarRenderPre(this.field_184059_f.field_195558_d, bossinfoclient, k, j, 10 + this.field_184059_f.field_71466_p.field_78288_b);
++                    net.minecraftforge.client.event.ForgeClientEventFactory.bossBarRenderPre(this.field_184059_f.field_195558_d, bossinfoclient, k, j, 10 + this.field_184059_f.field_71466_p.field_78288_b);
 +            if (!event.isCanceled()) {
              GlStateManager.func_179131_c(1.0F, 1.0F, 1.0F, 1.0F);
              this.field_184059_f.func_110434_K().func_110577_a(field_184058_a);
@@ -15,7 +15,7 @@
 -            j += 10 + this.field_184059_f.field_71466_p.field_78288_b;
 +            }
 +            j += event.getIncrement();
-+            net.minecraftforge.client.ForgeHooksClient.bossBarRenderPost(this.field_184059_f.field_195558_d);
++            net.minecraftforge.client.event.ForgeClientEventFactory.bossBarRenderPost(this.field_184059_f.field_195558_d);
              if (j >= this.field_184059_f.field_195558_d.func_198087_p() / 3) {
                 break;
              }
