+++ patches/minecraft/net/minecraft/client/renderer/color/ItemColors.java.patch
@@ -14,7 +14,7 @@
        itemcolors.func_199877_a((p_210237_0_, p_210237_1_) -> {
           return p_210237_1_ == 0 ? -1 : ItemMap.func_190907_h(p_210237_0_);
        }, Items.field_151098_aY);
-+      net.minecraftforge.client.ForgeHooksClient.onItemColorsInit(itemcolors, p_186729_0_);
++      net.minecraftforge.client.event.ForgeClientEventFactory.onItemColorsInit(itemcolors, p_186729_0_);
        return itemcolors;
     }
