+++ patches/minecraft/net/minecraft/client/renderer/model/ModelManager.java.patch
@@ -8,7 +8,7 @@
 +      net.minecraftforge.client.model.ModelLoader modelbakery = new net.minecraftforge.client.model.ModelLoader(p_195410_1_, this.field_174956_b);
        this.field_174958_a = modelbakery.func_177570_a();
        this.field_174955_d = this.field_174958_a.get(ModelBakery.field_177604_a);
-+      net.minecraftforge.client.ForgeHooksClient.onModelBake(this, this.field_174958_a, modelbakery);
++      net.minecraftforge.client.event.ForgeClientEventFactory.onModelBake(this, this.field_174958_a, modelbakery);
        this.field_174957_c.func_178124_c();
     }
