+++ patches/minecraft/net/minecraft/entity/monster/EntityMagmaCube.java.patch
@@ -4,7 +4,7 @@
     protected void func_70664_aZ() {
        this.field_70181_x = (double)(0.42F + (float)this.func_70809_q() * 0.1F);
        this.field_70160_al = true;
-+      net.minecraftforge.common.ForgeHooks.onLivingJump(this);
++      net.minecraftforge.event.ForgeEventFactory.onLivingJump(this);
     }
  
     protected void func_180466_bG(Tag<Fluid> p_180466_1_) {
