+++ patches/minecraft/net/minecraft/server/management/PlayerInteractionManager.java.patch
@@ -30,7 +30,7 @@
  
     public void func_180784_a(BlockPos p_180784_1_, EnumFacing p_180784_2_) {
 +      double reachDist = field_73090_b.func_110148_a(EntityPlayer.REACH_DISTANCE).func_111126_e();
-+      net.minecraftforge.event.entity.player.PlayerInteractEvent.LeftClickBlock event = net.minecraftforge.common.ForgeHooks.onLeftClickBlock(field_73090_b, p_180784_1_, p_180784_2_, net.minecraftforge.common.ForgeHooks.rayTraceEyeHitVec(field_73090_b, reachDist + 1));
++      net.minecraftforge.event.entity.player.PlayerInteractEvent.LeftClickBlock event = net.minecraftforge.event.ForgeEventFactory.onLeftClickBlock(field_73090_b, p_180784_1_, p_180784_2_, net.minecraftforge.common.ForgeHooks.rayTraceEyeHitVec(field_73090_b, reachDist + 1));
 +      if (event.isCanceled()) { // Restore block and te data
 +         field_73090_b.field_71135_a.func_147359_a(new SPacketBlockChange(field_73092_a, p_180784_1_));
 +         field_73092_a.func_184138_a(p_180784_1_, field_73092_a.func_180495_p(p_180784_1_), field_73092_a.func_180495_p(p_180784_1_), 3);
@@ -158,7 +158,7 @@
        } else if (p_187250_1_.func_184811_cZ().func_185141_a(p_187250_3_.func_77973_b())) {
           return EnumActionResult.PASS;
        } else {
-+         EnumActionResult cancelResult = net.minecraftforge.common.ForgeHooks.onItemRightClick(p_187250_1_, p_187250_4_);
++         EnumActionResult cancelResult = net.minecraftforge.event.ForgeEventFactory.onItemRightClick(p_187250_1_, p_187250_4_);
 +         if (cancelResult != null) return cancelResult;
           int i = p_187250_3_.func_190916_E();
           int j = p_187250_3_.func_77952_i();
@@ -180,7 +180,7 @@
        } else {
 -         boolean flag = !p_187251_1_.func_184614_ca().func_190926_b() || !p_187251_1_.func_184592_cb().func_190926_b();
 +         double reachDist = p_187251_1_.func_110148_a(EntityPlayer.REACH_DISTANCE).func_111126_e();
-+         net.minecraftforge.event.entity.player.PlayerInteractEvent.RightClickBlock event = net.minecraftforge.common.ForgeHooks
++         net.minecraftforge.event.entity.player.PlayerInteractEvent.RightClickBlock event = net.minecraftforge.event.ForgeEventFactory
 +                 .onRightClickBlock(p_187251_1_, p_187251_4_, p_187251_5_, p_187251_6_, net.minecraftforge.common.ForgeHooks.rayTraceEyeHitVec(p_187251_1_, reachDist + 1));
 +         if (event.isCanceled()) return event.getCancellationResult();
 +
