+++ patches/minecraft/net/minecraft/world/World.java.patch
@@ -684,7 +684,7 @@
                 d2 *= MoreObjects.firstNonNull(p_184150_11_.apply(entityplayer1), 1.0D);
              }
  
-+            d2 = net.minecraftforge.common.ForgeHooks.getPlayerVisibilityDistance(entityplayer1, d2, p_184150_9_);
++            d2 = net.minecraftforge.event.ForgeEventFactory.getPlayerVisibilityDistance(entityplayer1, d2, p_184150_9_);
 +
              if ((p_184150_9_ < 0.0D || Math.abs(entityplayer1.field_70163_u - p_184150_3_) < p_184150_9_ * p_184150_9_) && (p_184150_7_ < 0.0D || d1 < d2 * d2) && (d0 == -1.0D || d1 < d0)) {
                 d0 = d1;
